+++ assets/index.html
@@ -1,12 +1,12 @@
-<input type="text" id="resource">
+<input id="resource" type="text">
 <button onclick="window.startWebrtc(true)"> WebRTC Start Whip</button>
 <button onclick="window.startWebrtc(false)"> WebRTC Start Whep</button>
 
-<br />
-Video<br />
-<div id="remoteVideos"></div> <br />
+<br/>
+Video<br/>
+<div id="remoteVideos"></div> <br/>
 
-Logs<br />
+Logs<br/>
 <div id="div"></div>
 
 <script>
@@ -32,13 +32,13 @@ Logs<br />
             uri = "/whep/endpoint/" + resource;
         }
         if (whip) {
-            const screenStream = await navigator.mediaDevices.getDisplayMedia({ video: true });
+            const screenStream = await navigator.mediaDevices.getDisplayMedia({video: true});
             screenStream.getTracks().forEach(track => {
                 track.id = "c1861bb6-f6fc-46ad-965a-e53f24403212";
                 pc.addTrack(track, screenStream);
             });
         } else {
-            pc.addTransceiver('video', { 'direction': 'recvonly' })
+            pc.addTransceiver('video', {'direction': 'recvonly'})
             pc.ontrack = function (event) {
                 var el = document.createElement(event.track.kind)
                 el.srcObject = event.streams[0]
